[TRTLLM-12950][fix] MegaMoE-CuteDSL vendored kernel: Int64 SF indexing, FC1 SF pad granularity
Two review findings inside the vendored mega_moe_nvfp4 kernel (to be
upstreamed to the kernel team's swapab branch as well):

- token_comm token-back push_sf: promote the fc2_output_sf pool row
  index (and the peer combine_sf_u8 row, for symmetry with the adjacent
  Int64 DATA/metadata paths) to Int64 -- the Int32 offset overflows the
  multi-GiB SF pool at large max_tokens_per_rank x EP.
- kernel_fc12 FC1 weight-SF constexpr check: pad the M side to the
  128-row SF atom (as the FC2 path and the host packer already do)
  instead of the K-side sf_vec_size*4 granularity; behavior-identical
  for 128-aligned shapes like DSv4-Pro.

+++ tensorrt_llm/_torch/cute_dsl_kernels/mega_moe_nvfp4/token_comm.py
@@ -1324,10 +1324,14 @@ class TokenInPullTokenBackPush:
                             )
 
                 if cutlass.const_expr(self.push_sf):
-                    sf_local_addr = fc2_output_sf[pool_token_idx, 0,
+                    # Int64 like the DATA/metadata paths above: the Int32
+                    # row index would overflow the multi-GiB SF pool offset
+                    # at large max_tokens_per_rank x EP.
+                    sf_local_addr = fc2_output_sf[Int64(pool_token_idx), 0,
                                                   None].iterator.toint()
                     sf_peer_ptr = peer_rank_ptr_mapper.ptr_map_to_rank(
-                        combine_sf_u8[src_token, src_topk, None].iterator,
+                        combine_sf_u8[Int64(src_token), src_topk,
+                                      None].iterator,
                         src_rank,
                     )
                     for chunk in cutlass.range(num_sf_chunks, unroll=1):
